Separate the quantities for comparison in a file
