+++ 3-analyze/outputs/generate_plots.py
@@ -7,6 +7,8 @@ import numpy as np
 import pylab as pl
 import tqdm
 
+from quantities_for_comparison import birch_murnaghan
+
 def get_plugin_name():
     file_name = os.path.join(
         os.path.dirname(os.path.realpath(__file__)),
@@ -40,14 +42,6 @@ def get_conf_nice(configuration_string):
             ret_pieces.append(char)
     return "".join(ret_pieces)
 
-
-def birch_murnaghan(V,E0,V0,B0,B01):
-    r = (V0/V)**(2./3.)
-    return (E0 +
-            9./16. * B0 * V0 * (
-            (r-1.)**3 * B01 + 
-            (r-1.)**2 * (6. - 4.* r)))
-   
 
 if __name__ == "__main__":
     try:
